Remove d-objfile.cc, move parts to either decl or codegen
