+++ gcc/d/d-attribs.c
@@ -37,7 +37,7 @@ along with GCC; see the file COPYING3.  If not see
 #include "varasm.h"
 
 #include "d-tree.h"
-#include "d-objfile.h"
+#include "d-codegen.h"
 
 /* Internal attribute handlers for built-in functions.  */
 static tree handle_noreturn_attribute (tree *, tree, tree, int, bool *);
