+++ gcc/d/d-builtins.cc
@@ -35,7 +35,6 @@ along with GCC; see the file COPYING3.  If not see
 
 #include "d-tree.h"
 #include "d-codegen.h"
-#include "d-objfile.h"
 #include "id.h"
