+++ gcc/d/d-frontend.cc
@@ -35,7 +35,6 @@ along with GCC; see the file COPYING3.  If not see
 #include "stor-layout.h"
 
 #include "d-tree.h"
-#include "d-objfile.h"
 #include "d-codegen.h"
