+++ gcc/d/expr.cc
@@ -38,7 +38,6 @@ along with GCC; see the file COPYING3.  If not see
 
 #include "d-tree.h"
 #include "d-codegen.h"
-#include "d-objfile.h"
 #include "d-dmd-gcc.h"
 #include "id.h"
