+++ gcc/d/toir.cc
@@ -39,7 +39,6 @@ along with GCC; see the file COPYING3.  If not see
 
 #include "d-tree.h"
 #include "d-codegen.h"
-#include "d-objfile.h"
 #include "id.h"
